requirements.txt
absl-py==0.15.0
apache-beam==2.31.0
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
avro-python3==1.9.2.1
backcall==0.2.0
beautifulsoup4==4.8.0
blis==0.7.9
boto3==1.26.22
botocore==1.29.22
cachetools==4.2.4
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==14.0
confection==0.0.3
crcmod==1.7
cwl-eval==1.0.12
cymem==2.0.7
Cython==0.29.32
debugpy==1.6.4
decorator==5.1.1
dill==0.3.1.1
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
faiss-cpu==1.7.3
fastavro==1.7.0
filelock==3.8.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gevent==22.10.2
google==3.0.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grequests==0.6.0
grpcio==1.34.1
grpcio-tools==1.48.2
h5py==3.1.0
hdfs==2.7.0
httplib2==0.19.1
huggingface-hub==0.0.8
humanfriendly==10.0
idna==3.4
importlib-metadata==5.1.0
ipykernel==6.17.1
ipython==8.7.0
ipywidgets==8.0.2
ir-measures==0.3.1
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonlines==3.1.0
jupyter-client==7.4.8
jupyter-core==5.1.0
jupyterlab-widgets==3.0.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
langcodes==3.3.0
lightgbm==3.3.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
murmurhash==1.0.9
nest-asyncio==1.5.6
nltk==3.7
nmslib==2.1.1
numpy==1.19.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauth2client==4.1.3
oauthlib==3.2.2
onnxruntime==1.10.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathy==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
preshed==3.0.8
prompt-toolkit==3.0.33
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyathena==2.3.1
pybind11==2.6.1
pydantic==1.7.4
pydot==1.4.2
Pygments==2.13.0
pyjnius==1.4.2
pymongo==3.13.0
pyparsing==2.4.7
pyserini==0.16.0
pystream-protobuf==1.6.3
python-dateutil==2.8.2
pytrec-eval-terrier==0.5.5
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.5.4
sentence-transformers==2.0.0
sentencepiece==0.1.95
six==1.15.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
stack-data==0.6.2
tenacity==8.1.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
thinc==8.1.5
threadpoolctl==3.1.0
tokenizers==0.10.2
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.56.0
traitlets==5.6.0
transformers==4.6.1
typer==0.7.0
typing-extensions==3.7.4.3
urllib3==1.26.13
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.12.1
zipp==3.11.0
zope.event==4.5.0
zope.interface==5.5.2
textblob==0.17.1
pyqtgraph